Added module exports and require to files
