+++ app.js
@@ -16,6 +16,10 @@ app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
 app.use(express.static('public'))
 
+// routers
+const userRoute = require('./routes/userRoute')
+app.use('/', userRoute)
+
 // app.get('/', async (req, res) => {
 //   messageBoard
 //     .find()
