+++ routes/userRoute.js
@@ -1,27 +1,10 @@
 const express = require('express')
+const user = require('../controllers/userController')
 
 const userRouter = express.Router()
 
-userRouter
-  .route('/login')
-  .get(async (req, res) => {
-    res.render('login.ejs')
-  })
-  .post((req, res) => {
-    const { userName, password } = req.body
-    users
-      .findOne({ userName, password })
-      .then((result) => console.log('user found'))
-      .catch((err) => console.log('login error: ', err))
-    res.redirect('/')
-  })
+userRouter.route('/login').get(user.signIndex).post(user.signIn)
 
-userRouter.route('/register').post((req, res) => {
-  const { userName, password } = req.body
-  const user = {
-    userName,
-    password,
-  }
-  users.insertOne(user).then((result) => console.log(result))
-  res.render('login.ejs')
-})
+userRouter.route('/register').post(user.signUp)
+
+module.exports = userRouter
